Convert day 1 to aoc runner
