+++ input/2019/day1.txt
@@ -1,4 +1,4 @@
-pub const INPUT: &str = r#"94794
+94794
 58062
 112067
 139512
@@ -97,4 +97,4 @@ pub const INPUT: &str = r#"94794
 52801
 100693
 79229
-90365"#;
+90365
