+++ src/days/day2/input.rs
