+++ src/days/day3/map/cell.rs
