+++ src/days/day3/wire/direction.rs
