+++ src/days/day3/wire/path.rs
