+++ src/lib.rs
@@ -1,0 +1,8 @@
+extern crate aoc_runner;
+
+#[macro_use]
+extern crate aoc_runner_derive;
+
+pub mod days;
+
+aoc_lib! { year = 2019 }
